+++ CHANGELOG.md
@@ -5,6 +5,7 @@
      server descriptor doesn't contain such a summary anymore.
    - Remove optional fields "countries", "transports", and "versions"
      from clients objects which were still labeled as beta.
+   - Add new "version" parameter to filter for Tor version.
 
 
 # Changes in version 4.0-1.3.0 - 2017-08-04